1.3
Fixed issue with tiny blocks, the holder now is an ArmorStand just like normal and small ones.
Added more checks to every command and actions to avoid issues.
Added notifications when the blocks exist in config but cannot be found in the world.
Fixed issue with locations not updating correctly after editing a block's location.
Fixed issue with string which were being registered wrongly.